requirements.txt
# --- Core Web & API ---
fastapi==0.120.0
uvicorn[standard]==0.38.0
python-multipart
starlette==0.48.0
Jinja2==3.1.6
python-dotenv==1.1.1
itsdangerous

# --- Data Science ---
pandas
joblib

# --- ML Framework ---
tensorflow-cpu
scikit-learn

# --- Data Visualization ---
plotly==6.3.1

# --- Database & Connectors ---
mysql-connector-python==9.5.0

# --- Required Utilities ---
requests==2.32.5
httpx

# --- Security/Authentication ---
Authlib==1.6.5
werkzeug
google-generativeai